Fix menu list input press type

+++ src/menu_lst.cpp
@@ -222,7 +222,7 @@ int CMenuPcs::MLstCtrl()
 {
 	bool blocked;
 	float one;
-	unsigned int press;
+	u16 press;
 	u16 hold;
 	unsigned int itemCount;
 	unsigned int chunkCount;
